requirements.txt
absl-py==2.1.0
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
cryptography==41.0.7
filelock==3.13.1
fsspec==2023.12.2
gmpy2==2.1.2
google-auth==2.26.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
idna==3.4
Jinja2==3.1.2
kornia==0.7.1
lpips==0.1.4
Markdown==3.5.2
MarkupSafe==2.1.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.26.3
nvdiffrast==0.3.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
pip==23.3.1
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.4
setuptools==68.2.2
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
urllib3==1.26.18
Werkzeug==3.0.1
wheel==0.41.2
